refactor: split org_service into orgs/service, admin/service, users/service
- app/orgs/service.py: member CRUD (list, get, find, create, update, delete)
- app/admin/service.py: system-admin ops (list_all_orgs, rename, delete, transfer, list_all_users)
- app/users/service.py: user lifecycle (get_by_id, find_by_email, placeholder, get_or_create, delete)
- services/org_service.py + services/user_service.py become re-export shims
- Update orgs/routes.py and invite_service.py to import from domain modules
- Update auth/deps.py to import from users.service
- Update all test patch targets to new module paths
