+++ tests/unit/test_invite_service.py
@@ -34,12 +34,13 @@ def _mock_membership(is_active=True):
 
 
 @patch("app.services.invite_service.org_service")
-async def test_existing_user_gets_active_membership(mock_org_svc):
+@patch("app.services.invite_service.user_service")
+async def test_existing_user_gets_active_membership(mock_user_svc, mock_org_svc):
     """Known user who isn't a member → active membership, no Cognito call."""
     user = _mock_user()
     new_membership = _mock_membership()
 
-    mock_org_svc.find_user_by_email = AsyncMock(return_value=user)
+    mock_user_svc.find_user_by_email = AsyncMock(return_value=user)
     mock_org_svc.find_existing_membership = AsyncMock(return_value=None)
     mock_org_svc.create_membership = AsyncMock(return_value=new_membership)
 
@@ -67,12 +68,13 @@ async def test_existing_user_gets_active_membership(mock_org_svc):
 
 
 @patch("app.services.invite_service.org_service")
-async def test_active_member_raises_already_exists(mock_org_svc):
+@patch("app.services.invite_service.user_service")
+async def test_active_member_raises_already_exists(mock_user_svc, mock_org_svc):
     """User who is already an active member → AlreadyExists."""
     user = _mock_user()
     existing = _mock_membership(is_active=True)
 
-    mock_org_svc.find_user_by_email = AsyncMock(return_value=user)
+    mock_user_svc.find_user_by_email = AsyncMock(return_value=user)
     mock_org_svc.find_existing_membership = AsyncMock(return_value=existing)
 
     with pytest.raises(AlreadyExists, match="already a member"):
@@ -85,12 +87,13 @@ async def test_active_member_raises_already_exists(mock_org_svc):
 
 
 @patch("app.services.invite_service.org_service")
-async def test_pending_invite_raises_already_exists(mock_org_svc):
+@patch("app.services.invite_service.user_service")
+async def test_pending_invite_raises_already_exists(mock_user_svc, mock_org_svc):
     """User with a pending (inactive) invite → AlreadyExists."""
     user = _mock_user()
     pending = _mock_membership(is_active=False)
 
-    mock_org_svc.find_user_by_email = AsyncMock(return_value=user)
+    mock_user_svc.find_user_by_email = AsyncMock(return_value=user)
     mock_org_svc.find_existing_membership = AsyncMock(return_value=pending)
 
     with pytest.raises(AlreadyExists, match="already been invited"):
@@ -107,15 +110,16 @@ async def test_pending_invite_raises_already_exists(mock_org_svc):
 
 @patch("app.services.invite_service.invite_cognito_user")
 @patch("app.services.invite_service.org_service")
+@patch("app.services.invite_service.user_service")
 async def test_unknown_email_creates_cognito_and_placeholder(
-    mock_org_svc, mock_cognito
+    mock_user_svc, mock_org_svc, mock_cognito
 ):
     """Unknown email → Cognito invite + placeholder user + inactive membership."""
     placeholder = _mock_user(email="new@test.com")
     new_membership = _mock_membership(is_active=False)
 
-    mock_org_svc.find_user_by_email = AsyncMock(return_value=None)
-    mock_org_svc.create_placeholder_user = AsyncMock(return_value=placeholder)
+    mock_user_svc.find_user_by_email = AsyncMock(return_value=None)
+    mock_user_svc.create_placeholder_user = AsyncMock(return_value=placeholder)
     mock_org_svc.create_membership = AsyncMock(return_value=new_membership)
 
     db = _mock_db()
@@ -130,7 +134,7 @@ async def test_unknown_email_creates_cognito_and_placeholder(
 
     assert result == new_membership
     mock_cognito.assert_called_once_with("new@test.com")
-    mock_org_svc.create_placeholder_user.assert_called_once_with(db, "new@test.com")
+    mock_user_svc.create_placeholder_user.assert_called_once_with(db, "new@test.com")
     mock_org_svc.create_membership.assert_called_once_with(
         db=db,
         org_id=org_id,
@@ -144,12 +148,13 @@ async def test_unknown_email_creates_cognito_and_placeholder(
 
 
 @patch("app.services.invite_service.org_service")
-async def test_is_new_org_skips_duplicate_check(mock_org_svc):
+@patch("app.services.invite_service.user_service")
+async def test_is_new_org_skips_duplicate_check(mock_user_svc, mock_org_svc):
     """When is_new_org=True, don't bother checking for existing membership."""
     user = _mock_user()
     new_membership = _mock_membership()
 
-    mock_org_svc.find_user_by_email = AsyncMock(return_value=user)
+    mock_user_svc.find_user_by_email = AsyncMock(return_value=user)
     mock_org_svc.create_membership = AsyncMock(return_value=new_membership)
 
     await invite_user_to_org(
